Update proejct : Move base directory to ui
