+++ app/src/main/java/com/gdsc_cau/vridge/data/models/Voice.kt
@@ -1,4 +1,4 @@
-package com.gdsc_cau.vridge.models
+package com.gdsc_cau.vridge.data.models
 
 data class Voice(
     val data: String,
