+++ app/src/main/java/com/gdsc_cau/vridge/ui/profile/ProfileScreen.kt
@@ -1,4 +1,4 @@
-package com.gdsc_cau.vridge.profile
+package com.gdsc_cau.vridge.ui.profile
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
