+++ app/src/main/java/com/gdsc_cau/vridge/ui/talk/TalkScreen.kt
@@ -1,4 +1,4 @@
-package com.gdsc_cau.vridge.talk
+package com.gdsc_cau.vridge.ui.talk
 
 import androidx.compose.foundation.ScrollState
 import androidx.compose.foundation.background
