+++ app/src/main/java/com/gdsc_cau/vridge/ui/voicelist/VoiceListScreen.kt
@@ -1,10 +1,10 @@
-package com.gdsc_cau.vridge.voicelist
+package com.gdsc_cau.vridge.ui.voicelist
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.material3.Button
 import androidx.compose.runtime.Composable
-import com.gdsc_cau.vridge.main.Greeting
+import com.gdsc_cau.vridge.ui.main.Greeting
 
 @Composable
 fun VoiceListScreen(
